x/cfeairdrop/client/cli/query_user_airdrop_entries.go
package cli

import (
	"context"

	"github.com/chain4energy/c4e-chain/x/cfeairdrop/types"
	"github.com/cosmos/cosmos-sdk/client"
	"github.com/cosmos/cosmos-sdk/client/flags"
	"github.com/spf13/cobra"
)

func CmdListUserAirdropEntries() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "list-claim-record",
		Short: "list all userAirdropEntries",
		RunE: func(cmd *cobra.Command, args []string) error {
			clientCtx := client.GetClientContextFromCmd(cmd)

			pageReq, err := client.ReadPageRequest(cmd.Flags())
			if err != nil {
				return err
			}

			queryClient := types.NewQueryClient(clientCtx)

			params := &types.QueryUsersAirdropEntriesRequest{
				Pagination: pageReq,
			}

			res, err := queryClient.UsersAirdropEntries(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}

func CmdShowUserAirdropEntries() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "show-claim-record [address]",
		Short: "shows a userAirdropEntries",
		Args:  cobra.ExactArgs(1),
		RunE: func(cmd *cobra.Command, args []string) (err error) {
			clientCtx := client.GetClientContextFromCmd(cmd)

			queryClient := types.NewQueryClient(clientCtx)

			argIndex := args[0]

			params := &types.QueryUserAirdropEntriesRequest{
				Address: argIndex,
			}

			res, err := queryClient.UserAirdropEntries(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}
